chore(storybook): split shared config helpers
